refactor: 필터, AOP 수정
