+++ src/test/java/com/moabam/support/common/TestClockHolder.java
@@ -1,4 +1,4 @@
-package com.moabam.global.common.util;
+package com.moabam.support.common;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -6,17 +6,19 @@ import java.time.LocalDateTime;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
+import com.moabam.global.common.util.ClockHolder;
+
 @Component
-@Profile({"dev", "prod"})
-public class SystemClockHolder implements ClockHolder {
+@Profile("test")
+public class TestClockHolder implements ClockHolder {
 
 	@Override
 	public LocalDateTime times() {
-		return LocalDateTime.now();
+		return LocalDateTime.of(2023, 12, 3, 14, 30, 0);
 	}
 
 	@Override
 	public LocalDate date() {
-		return LocalDate.now();
+		return times().toLocalDate();
 	}
 }
